# api/openapi.yaml
openapi: 3.1.0
info:
  title: PVE-Longxin64 Control Plane API
  version: 0.1.0
  description: |
    PVE 云教室控制平面的首个纵切片契约。
    当前契约覆盖健康检查、总览、教室读取和教室批量任务创建。
    浏览器和瘦客户机不得直接访问 PVE API。
servers:
  - url: /api/v1
    description: Same-origin API
security: []
...
paths:
...
  /readiness:
    get:
      tags: [System]
      operationId: getReadiness
      summary: 检查 API 依赖是否就绪
      responses:
        "200":
          description: API 和存储已就绪
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessResponse"
        "503":
          description: 存储或必要依赖不可用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessResponse"
  /dashboard:
    get:
      tags: [Dashboard]
      operationId: getDashboard
      summary: 获取云教室总览
      responses:
        "200":
          description: 总览快照
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
        "500":
          $ref: "#/components/responses/InternalError"
  /classrooms:
    get:
      tags: [Classrooms]
      operationId: listClassrooms
      summary: 获取教室列表
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ClassroomStatus"
        - name: query
          in: query
          required: false
          description: 按教室名称或站点搜索
          schema:
            type: string
            maxLength: 100
      responses:
        "200":
          description: 教室列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassroomList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /classrooms/{classroomId}:
    get:
      tags: [Classrooms]
      operationId: getClassroom
      summary: 获取教室及座位状态
      parameters:
        - $ref: "#/components/parameters/ClassroomId"
      responses:
        "200":
          description: 教室详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassroomDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /classrooms/{classroomId}/operations:
    post:
      tags: [Classrooms, Operations]
      operationId: createClassroomOperation
      summary: 创建教室批量任务
      description: |
        只创建任务，不同步等待 PVE 操作完成。
        重复使用相同 Idempotency-Key 和相同请求时必须返回同一个任务。
      parameters:
        - $ref: "#/components/parameters/ClassroomId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassroomOperation"
      responses:
        "202":
          description: 任务已创建或返回已有幂等任务
          headers:
            Location:
              description: 任务查询地址
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 幂等键冲突或存在互斥任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          description: 目标范围或当前状态不允许执行
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          $ref: "#/components/responses/InternalError"
  /operations/{operationId}:
    get:
      tags: [Operations]
      operationId: getOperation
      summary: 获取批量任务及单项进度
      parameters:
        - $ref: "#/components/parameters/OperationId"
      responses:
        "200":
          description: 任务详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /operations/{operationId}/events:
    get:
      tags: [Operations]
      operationId: streamOperationEvents
      summary: 订阅批量任务增量事件
      description: |
        使用 Server-Sent Events 推送任务快照、父任务变化和单项变化。
        客户端断线后通过 `Last-Event-ID` 请求尚在有界历史中的事件；如果历史已淘汰、
        API 实例重启或请求来自另一实例，服务端会发送 `reset=true` 的完整快照。
        心跳使用 SSE 注释行，不占用事件序列；任务进入终态并发送完最终变化后流会正常结束。
        客户端必须保留轮询 `GET /operations/{operationId}` 作为降级路径。
      parameters:
        - $ref: "#/components/parameters/OperationId"
        - $ref: "#/components/parameters/LastEventId"
      responses:
        "200":
          description: |
            SSE 事件流。每个具名事件的 `data` 都是 OperationEvent JSON；
            `id` 与 JSON 中的 `sequence` 相同。
          headers:
            Cache-Control:
              description: 禁止缓存和代理转换事件流
              schema:
                type: string
            X-Accel-Buffering:
              description: 指示兼容反向代理不要缓冲事件
              schema:
                type: string
                enum: ["no"]
          content:
            text/event-stream:
              schema:
                oneOf:
                  - type: string
                    description: 实际 SSE 文本帧。
                  - type: array
                    description: 文档工具使用的逻辑事件序列；线上响应不会编码为 JSON 数组。
                    items:
                      $ref: "#/components/schemas/OperationEvent"
              example: |
                retry: 2000

                id: 42
                event: operation.updated
                data: {"event_type":"operation.updated","operation_id":"c7d188f0-18f3-4bf2-906d-e8f9ef66de8d","sequence":42,"timestamp":"2026-07-24T08:00:00Z","operation_status":"RUNNING","progress":{"total":50,"completed":10,"queued":30,"running":10,"succeeded":10,"failed":0,"skipped":0,"unknown":0},"resource_version":8}
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /operations:
    get:
      tags: [Operations]
      operationId: listOperations
      summary: 获取最近的批量任务
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: 任务列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  parameters:
    ClassroomId:
      name: classroomId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    OperationId:
      name: operationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    LastEventId:
      name: Last-Event-ID
      in: header
      required: false
      description: 最近已完整处理的事件序列；EventSource 重连时会自动发送。
      schema:
        type: integer
        format: int64
        minimum: 0
        maximum: 4503599627370000
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 客户端为一次业务意图生成的稳定 UUID。
      schema:
        type: string
        minLength: 1
        maxLength: 128
  responses:
    BadRequest:
      description: 请求格式错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    InternalError:
      description: 服务端错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    ReadinessResponse:
      type: object
      additionalProperties: false
      required: [status, checks, time]
      properties:
        status:
          type: string
          enum: [ready, not_ready]
        checks:
          type: object
          additionalProperties:
            type: string
        time:
          type: string
          format: date-time
    Dashboard:
      type: object
      additionalProperties: false
      required: [generated_at, summary, alerts]
      properties:
        generated_at:
          type: string
          format: date-time
        summary:
          $ref: "#/components/schemas/DashboardSummary"
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
    DashboardSummary:
      type: object
      additionalProperties: false
      required:
        - classrooms_total
        - classrooms_ready
        - classrooms_active
        - seats_total
        - seats_ready
        - thin_clients_online
        - desktops_running
        - operations_running
        - operations_failed
      properties:
        classrooms_total:
          type: integer
          minimum: 0
        classrooms_ready:
          type: integer
          minimum: 0
        classrooms_active:
          type: integer
          minimum: 0
        seats_total:
          type: integer
          minimum: 0
        seats_ready:
          type: integer
          minimum: 0
        thin_clients_online:
          type: integer
          minimum: 0
        desktops_running:
          type: integer
          minimum: 0
        operations_running:
          type: integer
          minimum: 0
        operations_failed:
          type: integer
          minimum: 0
    Alert:
      type: object
      additionalProperties: false
      required: [id, severity, title, description]
      properties:
        id:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        title:
          type: string
        description:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
        created_at:
          type: string
          format: date-time
    ClassroomList:
      type: object
      additionalProperties: false
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ClassroomSummary"
        total:
          type: integer
          minimum: 0
    ClassroomSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - site
        - status
        - timezone
        - seats_total
        - seats_ready
        - thin_clients_online
        - desktops_running
        - template_name
        - template_version
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        site:
          type: string
        status:
          $ref: "#/components/schemas/ClassroomStatus"
        timezone:
          type: string
          example: Asia/Shanghai
        seats_total:
          type: integer
          minimum: 0
        seats_ready:
          type: integer
          minimum: 0
        thin_clients_online:
          type: integer
          minimum: 0
        desktops_running:
          type: integer
          minimum: 0
        template_name:
          type: string
        template_version:
          type: string
        active_session:
          anyOf:
            - type: string
            - $ref: "#/components/schemas/ActiveSession"
            - type: "null"
        updated_at:
          type: string
          format: date-time
    ClassroomDetail:
      type: object
      additionalProperties: false
      required:
        - id
        - organization_id
        - name
        - site
        - status
        - timezone
        - template_name
        - template_version
        - resource_version
        - updated_at
        - seats
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        name:
          type: string
        site:
          type: string
        status:
          $ref: "#/components/schemas/ClassroomStatus"
        timezone:
          type: string
        template_name:
          type: string
        template_version:
          type: string
        active_session:
          type:
            - string
            - "null"
        resource_version:
          type: integer
          minimum: 1
        updated_at:
          type: string
          format: date-time
        seats:
          type: array
          items:
            $ref: "#/components/schemas/Seat"
    ClassroomStatus:
      type: string
      enum: [READY, ACTIVE, DEGRADED, OFFLINE]
...
    Seat:
      type: object
      additionalProperties: false
      required: [id, label, operation_state]
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        terminal:
          anyOf:
            - $ref: "#/components/schemas/ThinClient"
            - type: "null"
        desktop:
          anyOf:
            - $ref: "#/components/schemas/VirtualDesktop"
            - type: "null"
        operation_state:
          type: string
          enum: [QUEUED, RUNNING, WAITING_PVE, VERIFYING, SUCCEEDED, FAILED, SKIPPED, UNKNOWN]
        user_name:
          type:
            - string
            - "null"
...
    CreateClassroomOperation:
      type: object
      description: RESTORE 必须同时提供非空 reason 并将 confirmed 设为 true；其他操作不要求确认。
      additionalProperties: false
      required: [type]
      properties:
        type:
          $ref: "#/components/schemas/OperationType"
        seat_ids:
          type: array
          uniqueItems: true
          description: 目标座位；省略或传空数组表示整间教室，非空时仅操作列出的座位。
          items:
            type: string
            format: uuid
        reason:
          type: string
          maxLength: 500
          description: 操作原因；RESTORE 时必填且去除首尾空白后不能为空。
        confirmed:
          type: boolean
          default: false
          description: 破坏性操作确认；RESTORE 时必须为 true。
    Operation:
      type: object
      additionalProperties: false
      required:
        - id
        - classroom_id
        - type
        - status
        - counts
        - items
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        classroom_id:
          type: string
          format: uuid
        classroom_name:
          type: string
        type:
          $ref: "#/components/schemas/OperationType"
        status:
          $ref: "#/components/schemas/OperationStatus"
        reason:
          type: string
        request_id:
          type: string
        resource_version:
          type: integer
          minimum: 1
        counts:
          $ref: "#/components/schemas/OperationCounts"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OperationItem"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        started_at:
          type:
            - string
            - "null"
          format: date-time
        completed_at:
          type:
            - string
            - "null"
          format: date-time
    OperationItem:
      type: object
      additionalProperties: false
      required: [id, seat_id, seat_label, target_name, status, updated_at]
      properties:
        id:
          type: string
          format: uuid
        operation_id:
          type: string
          format: uuid
        seat_id:
          type: string
          format: uuid
        seat_label:
          type: string
        desktop_id:
          type:
            - string
            - "null"
          format: uuid
        cluster_id:
          type: string
          format: uuid
        pve_vmid:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [QUEUED, RUNNING, WAITING_PVE, VERIFYING, SUCCEEDED, FAILED, SKIPPED, UNKNOWN]
        target_name:
          type: string
        snapshot_name:
          type: string
          pattern: '^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$'
          description: RESTORE 任务创建时固化的基线快照名称；其他操作省略。
        upid:
          type: string
        error_code:
          type: string
        message:
          type: string
        started_at:
          type:
            - string
            - "null"
          format: date-time
        completed_at:
          type:
            - string
            - "null"
          format: date-time
        updated_at:
          type: string
          format: date-time
    OperationList:
      type: object
      additionalProperties: false
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Operation"
        total:
          type: integer
          minimum: 0
        generated_at:
          type: string
          format: date-time
    OperationEvent:
      type: object
      additionalProperties: false
      required:
        - event_type
        - operation_id
        - sequence
        - timestamp
        - operation_status
        - progress
        - resource_version
      properties:
        event_type:
          type: string
          enum: [operation.snapshot, operation.updated, operation.item.updated]
        operation_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        sequence:
          type: integer
          format: int64
          minimum: 1
          maximum: 9007199254740000
        timestamp:
          type: string
          format: date-time
        operation_status:
          $ref: "#/components/schemas/OperationStatus"
        item_status:
          type: string
          enum: [QUEUED, RUNNING, WAITING_PVE, VERIFYING, SUCCEEDED, FAILED, SKIPPED, UNKNOWN]
        seat_id:
          type: string
          format: uuid
        seat_label:
          type: string
        target_name:
          type: string
        error_code:
          type: string
        message:
          type: string
        item_updated_at:
          type: string
          format: date-time
        progress:
          $ref: "#/components/schemas/OperationProgress"
        resource_version:
          type: integer
          minimum: 1
        reset:
          type: boolean
          default: false
          description: 为 true 时客户端必须用本事件的 items 替换本地单项状态。
        items:
          type: array
          description: 仅快照事件携带的完整单项状态。
          items:
            $ref: "#/components/schemas/OperationItemEventState"
...
    OperationCounts:
      type: object
      additionalProperties: false
      required: [total, queued, running, succeeded, failed, skipped, unknown]
      properties:
        total:
          type: integer
          minimum: 0
        queued:
          type: integer
          minimum: 0
        running:
          type: integer
          minimum: 0
        succeeded:
          type: integer
          minimum: 0
        failed:
          type: integer
          minimum: 0
        skipped:
          type: integer
          minimum: 0
        unknown:
          type: integer
          minimum: 0
    OperationType:
      type: string
      enum: [PRECHECK, START, SHUTDOWN, RESTORE]
    OperationStatus:
      type: string
      enum:
        - QUEUED
        - VALIDATING
        - RUNNING
        - WAITING_PVE
        - VERIFYING
        - SUCCEEDED
        - PARTIALLY_SUCCEEDED
        - FAILED
        - CANCEL_REQUESTED
        - CANCELLED
    Problem:
      type: object
      additionalProperties: false
      required: [error_code, message, request_id]
      properties:
        error_code:
          type: string
        message:
          type: string
        request_id:
          type: string
        field_errors:
          type: object
          additionalProperties:
            type: string
        details:
          type: object
          additionalProperties: true